requirements.txt
graphviz==0.16
gensim==3.8.3
deepwalk==1.0.3
fasttext==0.9.2
xgboost==1.4.2
pyod==0.8.8
simplejson==3.16.0
scorecardpy==0.1.9.2
toad==0.1.0
pytz==2021.1
bayesian_optimization==1.2.0
pydantic==1.8.2
tensorflow==2.5.0
pypmml==0.9.11
jieba==0.42.1
sklearn2pmml==0.71.1
torch==1.7.1
mlxtend==0.18.0
category_encoders==2.2.2
matplotlib==3.4.2
scipy==1.6.3
tsfresh==0.18.0
pandas==1.2.4
gevent==21.1.2
requests==2.25.1
shap==0.39.0
lightgbm==3.2.1
sklearn_pandas==2.2.0
scikit-image==0.17.2
statsmodels==0.12.2
python_dateutil==2.8.1
node2vec==0.4.3
pyltp==0.2.1
openpyxl==3.0.10
networkx==2.8.2
numpy==1.19.5
scikit-learn==0.24.2
python-dateutil==2.8.1
nltk==3.7
textrank4zh==0.3